Fixed the FILE pointer issue
I wrote a lot of very detailed comments in the file and removed dead code (you don't need commented out code bcos you have git). You should go through all of the comments in detail. 
The main problem though was in readFromFile. Read my comments there and you will see. Basically, you need to pass in an fptr that has already been opened but instead, there was a LOCAL instance of fptr declared in the file.
We can a video conference to go over it. 
Read though it all and let it wash over you.